Allow arbitrary measures for cache size limit. Fixes #38.
This commit allows creating LruCaches that are limited in size by
something other than a count of entries. It does this by providing a
new trait, `Meter`, where implementations of the trait provide a
`measure` method to do the measurement. The default behavior of
measuring by the count of entries is provided by a `Count` type.

Due to difficulties with mutating entries making it possible to change
an item's measure, the `get_mut` and `iter_mut` methods are provided
only on LruCaches using `Count` for sizing.

A bit of optimization is provided for the default case--to avoid having
to redundantly store the item count the `Meter` trait allows implementations
to provide a type for storing the measurement, and the `Count` implementation
uses `()`. The `CountableMeter` trait allows `LruCache` to work with this
or any `Meter` implementation that uses `usize` for measurement.

Finally an optional feature is added, `heapsize`, which uses the heapsize
crate to provide HeapSizeOf, a measurement based on the heap memory usage
of values stored in the cache.